BF [2840359] - currency precision should be just an integer https://sourceforge.net/tracker/?func=detail&aid=2840359&group_id=176962&atid=879332

+++ base/src/org/compiere/model/I_I_PriceList.java
@@ -352,12 +352,12 @@ public interface I_I_PriceList
 	/** Set Price Precision.
 	  * Precision (number of decimals) for the Price
 	  */
-	public void setPricePrecision (BigDecimal PricePrecision);
+	public void setPricePrecision (int PricePrecision);
 
 	/** Get Price Precision.
 	  * Precision (number of decimals) for the Price
 	  */
-	public BigDecimal getPricePrecision();
+	public int getPricePrecision();
 
     /** Column name PriceStd */
     public static final String COLUMNNAME_PriceStd = "PriceStd";
